Quit if system doesn't support any sleep states
The motivation behind this is to help slow systems on battery power save
energy in cases where they do not support suspend but the distro/DE is
installing/starting sleep-inhibitor anyways.

This application can burn CPU cycles when the system is idle doing all
the various checks in the plugins. It's not a terrible amount of usage,
but every little bit helps.

In cases where sleep states are listed in sysfs, the app should continue
to work exactly as it did before.